Rename NewFeatureSupportLevel to FeatureSupportLevel
