+++ libs/ui-lib/lib/ocm/components/clusterConfiguration/operators/LvmCheckbox.tsx
@@ -17,7 +17,7 @@ import {
 import LvmHostRequirements from './LvmHostRequirements';
 import { OcmCheckboxField } from '../../ui/OcmFormFields';
 import { useTranslation } from '../../../../common/hooks/use-translation-wrapper';
-import { getLvmIncompatibleWithCnvReason } from '../../newFeatureSupportLevels/newFeatureStateUtils';
+import { getLvmIncompatibleWithCnvReason } from '../../featureSupportLevels/featureStateUtils';
 import { useNewFeatureSupportLevel } from '../../../../common/components/newFeatureSupportLevels';
 import NewFeatureSupportLevelBadge from '../../../../common/components/newFeatureSupportLevels/NewFeatureSupportLevelBadge';
