+++ libs/ui-lib/lib/ocm/components/index.ts
@@ -5,4 +5,4 @@ export * from './Router';
 export * from './AddHosts';
 export * from './clusterWizard';
 export * from './HostsClusterDetailTab';
-export * from './newFeatureSupportLevels';
+export * from './featureSupportLevels';
