--- libs/ui-lib/lib/ocm/components/newFeatureSupportLevels/index.ts
@@ -1,3 +1,0 @@
-export { default as NewFeatureSupportLevelProvider } from './NewFeatureSupportLevelProvider';
-export { default as ClusterNewFeatureSupportLevelsDetailItem } from './ReviewClusterNewFeatureSupportLevels';
-export * from './types';
